pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ResumeGPT"
version = "2.1"
description = "Tailor your resume to match any job posting effortlessly"
authors = [
    { name = "Tyler Kline", email = "tylerkline@gmail.com" }
]
maintainers = [
    { name = "Tyler Kline", email = "tylerkline@gmail.com" }
]
readme = "README.md"
license = { file = "LICENSE.txt" }
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "configparser>=7.1.0",
    "langchain>=0.2.11",
    "langchain-openai>=0.1.6",
    "langchain-core>=0.2.36",
    "langchain-community>=0.2.10",
    "pydantic>=2.7.1",
    "reportlab>=4.2.2",
    "requests>=2.32.3",
    "ruamel.yaml>=0.18.6",
    "pytest>=8.2.2",
    "python-dateutil>=2.8.0",
    "free-proxy>=1.1.1",
    "Jinja2>=3.1.4",
    "PyYAML>=5.3.1",
    "pytest-cov>=5.0.0"
]
keywords = ["resume", "job", "application", "automation", "job search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2"
]

[project.urls]
Homepage = "https://github.com/takline/ResumeGPT"
"Bug Tracker" = "https://github.com/takline/ResumeGPT/issues"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.github
  | \.pytest_cache
  | \.__pycache__
)/
'''
